--- app/src/main/java/com/kojingu/xylo/screens/MyViewModel.kt
@@ -1,10 +1,0 @@
-package com.kojingu.xylo.screens
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-
