+++ app/src/main/java/com/kojingu/xylo/screens/maingame/MainGameViewModel.kt
@@ -1,0 +1,4 @@
+package com.kojingu.xylo.screens.maingame
+
+class MainGameViewModel {
+}
